[Justice Counts] Add additional contexts box for all metrics, regardless of includes excludes (Recidiviz/recidiviz-data#25511)
## Description of the change

Matt and Stephanie noticed that for a metric that did not have any
includes/excludes (Jails Readmissions), there was no option to provide
additional contexts for that metric.
![Screenshot 2023-11-15 at 2 09
29 PM](https://github.com/Recidiviz/recidiviz-data/assets/82831800/c8056d2b-7a34-414f-bca8-bade78f403a7)

This PR changes this so that all metrics have an additional contexts
box, regardless of if they have includes/excludes or not.

How I Tested: I ran this branch locally and the Jails Readmissions
settings page looks as follows
![Screenshot 2023-11-15 at 2 12
15 PM](https://github.com/Recidiviz/recidiviz-data/assets/82831800/568f3c9f-3336-4ddb-bfcb-e0ec1ba6eeb5)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#25492

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@Michelles-MBP.attlocal.net>
GitOrigin-RevId: 6b7a00bc61233cf5e8f481086b2d8fd77b245dc7