[US_TN] Adding response_type CITATION for violations that are associated to ZTVR which are currently null response_types (Recidiviz/recidiviz-data#25510)
## Description of the change

In TN, they will have a number of violations that are associated to zero
tolerance violations on the same day. When this happens, TN does not
document the response for the violation with the SanctionStatus of
ASSOCIATED TO ZERO TOLERANCE and instead, only will document the
response to the main violation in that set of violations. Because of
this practice, all of the violations with SanctionStatus = ASSOCIATED TO
ZERO TOLERANCE have a null response_type currently. However, this does
not work for us downstream because the only way a violation will get in
included in someone's violation history window is if the response_type
is non-null.

This PR addresses this issue by adding in conditional logic in the
mapping for ViolationAndSanctions that will add an enum of CITATION to
any response with SanctionStatus = "Associated to Zero Tolerance" which
means these violations will now be included in violation_metrics for a
better understanding of violation history downstream.

This PR also includes a STATE migration to apply this mappings change in
the STATE db to avoid a rerun.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#24839

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4511fc17b141dd26166f825e1554caf08a376e7f