[US_TN] adding some validation bumps based on outliers work with TEMPORARY_CUSTODY  (Recidiviz/recidiviz-data#25559)
## Description of the change

When we decided to include TEMPORARY_CUSTODY periods in populations,
some of our external validations started hard failing higher than
before, so this PR adds slight bumps to our thresholds which we can
adjust again if we get feedback about their inclusion or not.

Additionally, Recidiviz/recidiviz-data#25390 updated some admission_reasons without changing the
PFI which caused PFI related validations to fail, so we add that in
which should resolve those validation errors.

Finally, we have some staff in TN with job titles of officers whose
location is not a supervision site. We confirmed using the active roster
from TN for supervisors that these people are not active supervision
officers with caseloads and therefore we bump the threshold to allow
some supervision officer roles for people not in supervision offices.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#24963, Recidiviz/recidiviz-data#24834, Recidiviz/recidiviz-data#24966, Recidiviz/recidiviz-data#24964, Recidiviz/recidiviz-data#248365, Recidiviz/recidiviz-data#24836

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6c12efc89d3406092030d6b4afd9a7d930e094b8