[Justice Counts] Surface child agency IDs when returning agency JSONs from admin panel (Recidiviz/recidiviz-data#25509)
## Description of the change

This PR modifies all agency API endpoints in the New Justice Counts
admin panel to include a list of child agency IDs when returning JSON
representations of agencies. If the requested agency is a superagency,
the response will include a list of its child agency IDs. If the agency
is not a superagency, the response will include an empty list.

## Related issues

Closes Recidiviz/recidiviz-data#25357

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: db92da260dc907ca6019ab441cf509770a38f360