[Justice Counts] add includes_excludes descriptions to all dimensions, regardless of includes/excludes (Recidiviz/recidiviz-data#25550)
## Description of the change

Related to Recidiviz/recidiviz-data#25511, if a dimension does not have any includes/excludes,
previously the user was not able to provide an includes_excludes
description for that dimensions. This PR changes this so that all
dimensions have an includes_excludes description box, regardless of if
they have includes/excludes or not. The only exception to this are
`OTHER` and `UKNOWN` dimensions, which have only additional contexts
descriptions.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#25511

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@Michelles-MBP.attlocal.net>
GitOrigin-RevId: 6f5706069f77f2a555029bc32b3d00685877bfdf