Nick Tallant IE Onboarding, Validation (Recidiviz/recidiviz-data#25577)
ONBOARDING Exercise: This adds the `US_OZ_LOTR_ID` external ID type for the lotr data
system in oz to the validation view query. The raw data was updated and
added to the staging ingest bucket, and I added what that data looked
like in the docstring for the next person to onboard.

GitOrigin-RevId: b42b6ade95253b2e8fc058863d571dfe877c8cb7