[US_PA] Update location reference files (Recidiviz/recidiviz-data#25608)
## Description of the change

In the process of refactoring `supervision_staff_location_period` view
in PA, I adjusted the way we were ingesting locations so that we can
simplify `us_pa_location_metadata` and the
`supervision_locations_from_contacts` reference table. Unfortunately, we
do still need to ingest a reference table of supervision location names
to location IDs for locations that come from supervision contacts data,
even though PA sends us this information directly from their system in
`dbo_LU_PBPP_Organization`, because the location names are spelled and
abbreviated differently in practice than they are in the
system-generated reference table.

I used that reference table to generate my own, which associates the
proper org codes with the location names that appear in the contacts
data. The location names from the roster we just received are formatted
correctly (that is, the same as in `dbo_LU_PBPP_Organization`), so the
combination of `supervision_locations_from_contacts` and
`dbo_LU_PBPP_Organization` will allow us to translate English location
names into their IDs in the ingest view.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1541614a4e9445874680562a5fb3654551f2093b