[US_TN] removing WRB from custom parser for sup_type in TN  (Recidiviz/recidiviz-data#25639)
## Description of the change

Related to Recidiviz/recidiviz-data#25627, we are removing the `WRB` mappings in the custom
parser in TN to remove mapping this value to BENCH_WARRANT to improve
downstream metrics.

Because of the nature of this custom parser, this change will require a
partial rerun of the SP view to apply this change properly.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#25627

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e4d70a7e679b472e63c80f4c5ff44350d03f1215