[Validation] Bugfixes / preparatory refactor to validate_state_dataset script (Recidiviz/recidiviz-data#25649)
## Description of the change

This PR does two things:
1) Refactors the `validate_state_dataset` code so that more of it
supports tables that don't directly map to an `Entity` subclass.
Introduces a `_materialize_comparable_table` which reduces some code
duplication, and which will pick the query for building a comparable
table based on whether it's an association table or not (association
table part not implemented yet).
2) Fixes a bug where we weren't actually pulling the standardized
`person_id`/`staff_id` columns into the "comparable" table, so we were
counting two entities as identical if they had all the same fields but
were attached to a different person. In practice, this won't have large
impact because most entities have an external_id that is unique to the
person, but it will impact entities like `StatePersonRace`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#24413.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 45e3236281d527c9d47e494a33a388ddc325594b