Support aggregated metrics for non-person units of observation (Recidiviz/recidiviz-data#24069)
## Description of the change

This change adds support for units of analysis and populations based on
other units of observation besides the `person_id`. For instance, this
allows us to configure a "district" unit of analysis where the spans of
assignment to the district are encoded at the officer level, rather than
at the person level (similarly for a population, where a supervision
population can be defined in terms of officer assignments to the
supervision population, rather than person assignments).

This also extends support for aggregated metrics to be configured at
alternate units of assignment, so that we can reference events and spans
for non-person levels of observation. This PR adds one aggregated metric
for workflows usage metrics at the officer level for the sake of
illustration.

This update will allow us to track usage metrics in aggregated metrics.

(See [design
doc](https://paper.dropbox.com/doc/WIP-Design-Doc-Extending-aggregated-metrics-infrastructure-to-usage-events--B~GPGvBlHvFtlCaQNKkkpx64Ag-O8wphbeCES6nTjzOekwmI))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#23893

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
Co-authored-by: Anna Geiduschek <anna@recidiviz.org>
GitOrigin-RevId: 655b21fb64de72ac5b415162c25f47ae6c58c4e9